refactor(jvm-utilities): the toObjectString() method changed its behavior again.
Previously, its behavior was based on searching for any toString method recursively and taking it for printing. The problem with this was that if this method was inside the same class it could make recursive calls (this method is meant to be called inside toString() methods) and fill the stack with calls.

Now an observeStack call has been generated and all it does is store all the calls that are made to a certain object. If the object makes many consecutive calls (more than 2 calls is already considered recursive), it blocks it and shows a generic response along with the error that there is a recursion inside the called method. In this way, frozen programs and some unexpected error are avoided.